Move Choices and Selections into the global scope
This makes it possible for the console to access these variables.
Otherwise there is no way for the console to access these variables. So
the round can only be played with multiple console.logs written into the
playRound function itself, which is not the intended use. Also the
playGame function would not work either.