Complete rock paper scissors game
In order to make the game play five rounds scope was important to keep
in mind. Playing one round was possible with functions in the scope of
the previous commit but in order to make new selections it needed to be
reworked. Also using function delarations rather than expressions helped
the final code look much neater to me.